some cleaning of code and used useNavigator hook to make home page link work
