+++ src/components/directory-item/directory-item.component.jsx
@@ -1,14 +1,17 @@
+import { useNavigate } from 'react-router-dom';
 import './directory-item.styles.scss';
 
 const DirectoryItem = ({ category }) => {
-  const { imageUrl, title } = category;
+  const { imageUrl, title, route} = category;
+
+  const navigate = useNavigate();
+
+  const onNavigateHandler = () => navigate(route)
   return (
-    <div className='directory-item-container'>
+    <div className='directory-item-container' onClick={onNavigateHandler}>
       <div
         className='background-image'
-        style={{
-          backgroundImage: `url(${imageUrl})`,
-        }}
+        style={{backgroundImage: `url(${imageUrl})`}}
       />
       <div className='body'>
         <h2>{title}</h2>
